Well it might be not an Responsive  for now But soon it will be
"Following my brother's idea, I am trying out your learning approach this time. I plan to implement it as well, and hopefully, I will make it responsive soon. Remember, you will learn not just the coding but also how to try. For now, it might not look perfect, like that of a pro developer, but it's a starting point for me. It might feel like a waste of time for some, but for me, it's the first step in this field. I will be making more like that soon, and they will be responsive as well.